+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 - [FSharp] Added non-generic isNull and notNull
 - [FSharp] Added generic inc and dec
+- [FSharp] Fixed generic constraint of NativePtr.stackUseArr
 
 ### 5.3.16
 - Fixed Cell.XYZ bug and improved doc comments