+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
+### 5.3.17
 - [FSharp] Added non-generic isNull and notNull
 - [FSharp] Added generic inc and dec
 - [FSharp] Fixed generic constraint of NativePtr.stackUseArr